requirements.txt
Flask
Werkzeug==2.3.6
openai==0.27.8
PyPDF2==3.0.1
python-docx==0.8.11
